+++ torch_geometric/datasets/__init__.py
@@ -115,6 +115,7 @@ from .explainer_dataset import ExplainerDataset
 from .infection_dataset import InfectionDataset
 from .ba2motif_dataset import BA2MotifDataset
 from .ba_multi_shapes import BAMultiShapesDataset
+from .rbgv import RBGVDataset
 from .ba_shapes import BAShapes
 
 import torch_geometric.datasets.utils
@@ -243,6 +244,7 @@ synthetic_datasets = [
     'InfectionDataset',
     'BA2MotifDataset',
     'BAMultiShapesDataset',
+    'RBGVDataset',
     'BAShapes',
 ]